\name{calc_icvi}
\alias{calc_icvi}
\title{Calculate Item-Content Validity Index (I-CVI)}
\usage{
calc_icvi(data)
}
\arguments{
\item{data}{A dataframe or matrix with items in rows and experts in columns.}
}
\value{
A numeric vector of I-CVI values for each item.
}
\description{
Calculate Item-Content Validity Index (I-CVI)
}
