% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations.R
\name{calc_kappa}
\alias{calc_kappa}
\title{Calculate Modified Kappa Statistic}
\usage{
calc_kappa(data)
}
\arguments{
\item{data}{A dataframe or matrix with items in rows and experts in columns.}
}
\value{
A numeric vector of the modified kappa statistic for each item.
}
\description{
Calculate Modified Kappa Statistic
}
