% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations.R
\name{calc_scvi_ave}
\alias{calc_scvi_ave}
\title{Calculate Scale-Content Validity Index/Average (S-CVI/Ave)}
\usage{
calc_scvi_ave(data)
}
\arguments{
\item{data}{A dataframe or matrix with items in rows and experts in columns.}
}
\value{
A single numeric value representing S-CVI/Ave.
}
\description{
Calculate Scale-Content Validity Index/Average (S-CVI/Ave)
}
