% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations.R
\name{calc_scvi_ua}
\alias{calc_scvi_ua}
\title{Calculate Scale-Content Validity Index/Universal Agreement (S-CVI/UA)}
\usage{
calc_scvi_ua(data)
}
\arguments{
\item{data}{A dataframe or matrix with items in rows and experts in columns.}
}
\value{
A single numeric value representing S-CVI/UA.
}
\description{
Calculate Scale-Content Validity Index/Universal Agreement (S-CVI/UA)
}
