% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations.R
\name{cvi_summary}
\alias{cvi_summary}
\title{CVI Summary}
\usage{
cvi_summary(data)
}
\arguments{
\item{data}{A dataframe or matrix with items in rows and experts in columns.}
}
\value{
A list containing I-CVI, S-CVI/Ave, S-CVI/UA, and Kappa.
}
\description{
CVI Summary
}
