% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_icvi}
\alias{plot_icvi}
\title{Plot I-CVI values}
\usage{
plot_icvi(data, threshold = 0.78)
}
\arguments{
\item{data}{A dataframe or matrix with items in rows and experts in columns.}

\item{threshold}{A numeric value for the CVI threshold (default 0.78).}
}
\value{
A ggplot2 bar plot of I-CVI values.
}
\description{
Plot I-CVI values
}
