% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_ratings}
\alias{plot_ratings}
\title{Plot Ratings Heatmap}
\usage{
plot_ratings(data)
}
\arguments{
\item{data}{A dataframe or matrix with items in rows and experts in columns.}
}
\value{
A ggplot2 tile plot of ratings.
}
\description{
Plot Ratings Heatmap
}
